% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{chkFit}
\alias{chkFit}
\title{(internal) Checking for the fit of the predictions}
\usage{
chkFit(df, sg, FYCFs, itrY, opt, title)
}
\arguments{
\item{df}{a data.frame}

\item{sg}{a string for the subgroup expression}

\item{FYCFs}{a list of glm results}

\item{itrY}{a vector of Y grids}

\item{opt}{a list of options}

\item{title}{a string, title}
}
\value{
FYPred which is the prediction of FY for each observation.
}
\description{
proc to check fit by ploting as well as returning FY predictions
It internally returns the smoothed predictions as well as plots. Plots are muted in default
}
