% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{displayMeans}
\alias{displayMeans}
\title{(internal) Displaying the mean estimation tables with CIs
generating table of mean estimates with CIs}
\usage{
displayMeans(repV, optset, n, repVLB = NA, curB,
  showDisplayMean = TRUE)
}
\arguments{
\item{repV}{reportingV object}

\item{optset}{a list of options}

\item{n}{the sample size}

\item{repVLB}{reportingV object for the partial identification. Default is NA}

\item{curB}{an integer, current value of the bootstrap procedure}

\item{showDisplayMean}{a boolean, if TRUE, then it shows an intermediary table}
}
\value{
None.
}
\description{
(internal) Displaying the mean estimation tables with CIs
generating table of mean estimates with CIs
}
