% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{genDR}
\alias{genDR}
\title{(internal) generating distribution regressions}
\usage{
genDR(df, sg, itr, opt, fml, link)
}
\arguments{
\item{df}{a data.frame}

\item{sg}{a string, expression for the subgroup}

\item{itr}{a vector, equispaced grids}

\item{opt}{a list containing the options}

\item{fml}{a formula object}

\item{link}{a string, link function}
}
\value{
FCFs, a list of glm results for every grids for Y to be evaluated.
}
\description{
distribution regression routine
}
