% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{genPlot2way}
\alias{genPlot2way}
\title{(internal) Two-way plot of two cdfs}
\usage{
genPlot2way(itr, F1, F2, label1, label2, title)
}
\arguments{
\item{itr}{a vector of grids for x axis}

\item{F1}{a vector of values for first objects for y axis}

\item{F2}{a vector of values for second objects for y axis}

\item{label1}{a string, label for F1}

\item{label2}{a string, label for F2}

\item{title}{a string, title for the plot}
}
\value{
ggplot object readily printable.
}
\description{
Function to return quantile for a given ranking
Returns list of functions of chebyshev interpolated distribution coefficients
}
