% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{getQuantile}
\alias{getQuantile}
\title{(internal) Take numerical inverse of the cdf}
\usage{
getQuantile(FY, kI, kU)
}
\arguments{
\item{FY}{a vector of values}

\item{kI}{a vector of grids for FY}

\item{kU}{a vector of grids to evaluate}
}
\value{
QY quantile of FY
}
\description{
get quantile by inversion
}
