% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{integrate1}
\alias{integrate1}
\title{(internal) Numerical integration with respect to the estimated cdf}
\usage{
integrate1(itr, FY, opt)
}
\arguments{
\item{itr}{grids}

\item{FY}{values}

\item{opt}{a list of options}
}
\value{
mean of integrated value with respect to the values FY
}
\description{
Numerical integration for usual grids
}
