% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{mainProcedure}
\alias{mainProcedure}
\title{(internal) Main caller for the estimation procedure}
\usage{
mainProcedure(main, optset, fml, Ytype, link)
}
\arguments{
\item{main}{data.frame of the reformatted dataset}

\item{optset}{a list of options}

\item{fml}{a formula reformatted ready for the glm}

\item{Ytype}{a string, either "Y1" or "Yb"}

\item{link}{a string, link function for the glm}
}
\value{
reporting of the list of mean estimates, $reportingV and $reportingQ
}
\description{
the common procedure called by both mainestimate and bootstrapProc
}
