% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{meanEstimate}
\alias{meanEstimate}
\title{Estimation of mean effects and quantile differences}
\usage{
fml <- "logoutput ~ head_age_bl + members_resid_bl | act_livestock_bl + act_business_bl + borrowed_total_bl + ccm_resp_activ + other_resp_activ + ccm_resp_activ_d + nadults_resid_bl"
meanEstimate(df=data,dDim=2,treatment="treatment",posttreat="client",numKnots=1,numOrder=1,formula=fml,Ytype="Yb")
# head_age_bl and members_resid_bl are converted to 2-dim B-spline, and the rest will be factorized.

fml <- "logoutput ~ head_age_bl + members_resid_bl + head_age_bl*members_resid_bl + act_livestock_bl + act_business_bl + borrowed_total_bl + ccm_resp_activ + other_resp_activ + ccm_resp_activ_d + nadults_resid_bl"
meanEstimate(df=data,dDim=2,treatment="treatment",posttreat="client",formula=fml,Ytype="Yb",typeSieve="polynomial")
# regard the formula give as the exact polynomial expression to be used for the estimations.
}
\arguments{
\item{df}{a data.frame, the exact dataset used for the study.}

\item{dDim}{an integer, positive, the number of values D would take}

\item{treatment}{a string, variable name for the treatment variable.}

\item{posttreat}{a string, variable name for the post-treatment variable.}

\item{typeSieve}{a string, either "formula" or "spline". Default is "spline" taking B-spline for the continuous variables specified in formula}

\item{formula}{a formula object, for "spline", specify as "outcome ~ continuous | discrete". For "polynomial", use as in other packages like "lm". See the examples for detailed uses.}

\item{numOrder}{(for typeSieve = "spline") an integer, positive, the maximum order of the spline polynomial}

\item{numKnots}{(for typeSieve = "spline") an integer, positive, the number of knots for the spline}

\item{dumAllAdditive}{(for typeSieve = "spline") a boolean, if TRUE, then the dummies are additively included. Default is FALSE and splines are allowed different across discrete variables term.}

\item{additiveSpline}{(for typeSieve = "spline") a boolean, if TRUE, then the splines are additively combined. Default is FALSE and uses the kronecker product of uni-dimensional splines.}

\item{nGrids}{an integer, positive, a number of grids for the outcome in the estimation for the conditional cdfs. Default is set to 30. This number crucially affects the computation time as it requires this many of glm estimations separately.}

\item{nGridsFine}{an integer, positive, a number of grids used for the numerical integration. Default is set to 1000.}

\item{plotBeg}{a float, a value in (0,1), the lower range of the quantile difference estimate.}

\item{plotEnd}{a float, a value in (0,1), the upper range of the quantile difference estimate.}

\item{plotBy}{a float, a positive value, the increment for the quantile difference esimate.}

\item{link}{a string, a link function used as a binomial family link for glm. Refer to the specification of the link options for the R package glm. Defalt is "logit"}

\item{optMute}{a boolean, option to toggle off the outputs used during the development. Default is TRUE.}

\item{Ytype}{a string, either "Yb" or "Y1". Default is "Yb" which uses a proxy variable for Y0. Specification with "Y1" uses Y1 as a proxy for Y0.}

\item{discY}{a boolean, option to toggle on an experimental feature for the partial identification with discrete proxy variable. Default is FALSE, which assumes absolutely continuous Yb.}
}
\value{
returns a list, reporting, containing a vector of average effects $reportingV, and a vector of quantile differences $reportingQ.
}
\description{
This function returns the mean estimates of the average effects and the quantile effects.
}
\section{Note}{

typeSieve = "formula" uses the argument formula as a specification of the linear sieve.
}

