% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{plotQuantile}
\alias{plotQuantile}
\title{(internal) Generating and saving the plots of the quantiles with uniform CIs
generating plot of quantiles with uniform CI}
\usage{
plotQuantile(lenQ, Qmat, optset, alpha, n, filename)
}
\arguments{
\item{lenQ}{an integer, length of Q}

\item{Qmat}{a matrix, matrix of bootstrapped quantile estimates}

\item{optset}{a list of options}

\item{alpha}{a float, the size of the test}

\item{n}{an integer, sample size}

\item{filename}{a string for the path to save the plot figures}
}
\value{
none.
}
\description{
(internal) Generating and saving the plots of the quantiles with uniform CIs
generating plot of quantiles with uniform CI
}
