% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{predFY}
\alias{predFY}
\title{(internal) Prediction for the conditional cdfs}
\usage{
predFY(df, sg, FYCFs, itr, opt, n)
}
\arguments{
\item{df}{data.frame}

\item{sg}{a string, expression for the subgroup}

\item{FYCFs}{a glm results, distribution regression coefficients}

\item{itr}{a vector, grids}

\item{opt}{a list, options}

\item{n}{an integer, subsample size}
}
\value{
FYWpred, prediction of FY for each observation
}
\description{
prediction of FY conditional on W
}
