% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{reformatFormula}
\alias{reformatFormula}
\title{(internal) Transoformation of the data.frame according to the formula specification}
\usage{
reformatFormula(fml, dataOrigin, treatment, posttreat, cluster = "")
}
\arguments{
\item{fml}{a formula object}

\item{dataOrigin}{a data.frame object, original dataset}

\item{treatment}{a string, containing treatment variable varname string}

\item{posttreat}{a string, containing post-treatment variable varname string}

\item{cluster}{a string, containing cluster variable varname. Default is "" which means either its mean process or not clustered.}
}
\value{
a data.frame which contains covariates W, treatment T, posttreat D, and the outcome Y as well as the cluster.
}
\description{
This is an internal function extracting the variable names from the formula to modify the data.frame given
so that the formatted variable names are used in the code inside.
}
