% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{retFormula}
\alias{retFormula}
\title{(internal) returns formula according to the spline specification}
\usage{
retFormula(opt)
}
\arguments{
\item{opt}{a list containing options}
}
\value{
a formula, fml
}
\description{
return formula expressions for spline basis used in distribution regressions
}
