% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procFunctions.R
\name{showDiff}
\alias{showDiff}
\title{(internal) Showing the treatment effects}
\usage{
showDiff(df, Pred1, Pred0, seq1, seq0, itr, opt)
}
\arguments{
\item{df}{data.frame}

\item{Pred1}{prediction of Y1}

\item{Pred0}{prediction of Y0}

\item{seq1}{a string, subgroup expression for Y1}

\item{seq0}{a string, subgroup expression for Y0}

\item{itr}{grids for Y}

\item{opt}{a list of options}
}
\value{
None.
}
\description{
Return Treatment Effects
}
\details{
Muted in default.
}
